# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDomain API
  description: Finally a simple interface for domain names.
  version: "0.1.0"
  contact:
    email: contact+api@happydomain.org
    name: happyDomain team
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
...
servers:

  - url: http://localhost:8081/api
    description: Development server

  - url: https://app.happydomain.org/api
    description: Main production server
paths:
...
  /api/providers:

    get:
      operationId: get_providers
      description: Retrieve all DNS providers belonging to the user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas.yaml#/components/schemas/ProviderMeta'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Unable to retrieve user's domains
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Retrieve user's providers
      tags:
        - providers

    post:
      operationId: create_provider
      description: Append a new provider for the user.
      requestBody:
        description: Provider to add
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/ProviderMinimal'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "400":
          description: Error in received data
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Unable to retrieve current user's providers
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Add a new provider
      tags:
        - providers
...
  /api/providers/_specs/{providerType}/settings:

    post:
      operationId: create_provider_from_settings
      description: This creates or updates a Provider with human fillable forms.
      parameters:
      - description: The provider's type
        in: path
        name: providerType
        required: true
        schema:
          type: string
      requestBody:
        description: The current state of the Provider's settings, possibly empty
          (but not null)
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/ProviderSettingsState'
      responses:
        "200":
          description: The Provider has been created with the given settings
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "202":
          description: The settings need more refinement
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ProviderSettingsResponse'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Assistant to Provider creation.
      tags:
        - provider_specs
  /api/providers/{providerId}:

    delete:
      operationId: delete_provider
      description: Delete a Provider from the database. It is required that no Domain
        are still managed by this Provider before calling this route.
      parameters:
      - description: Provider identifier
        in: path
        name: providerId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Provider deleted
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Database writing error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Delete a Provider.
      tags:
        - providers

    get:
      operationId: get_provider
      description: Retrieve information in the database about a given Provider owned
        by the user.
      parameters:
      - description: Provider identifier
        in: path
        name: providerId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Retrieve Provider information.
      tags:
        - providers

    put:
      operationId: update_provider
      description: Updates the information about a given Provider owned by the user.
      parameters:
      - description: Provider identifier
        in: path
        name: providerId
        required: true
        schema:
          type: string
      requestBody:
        description: The new object overriding the current provider
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/Provider'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "400":
          description: Identifier changed
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Database writing error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Update Provider information.
      tags:
        - providers
  /api/providers/{providerId}/domains:

    get:
      operationId: list_provider_domains
      description: List domains available from the given Provider.
      parameters:
      - description: Provider identifier
        in: path
        name: providerId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "400":
          description: Provider error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Lists manageable domains from the Provider.
      tags:
        - providers
  /api/providers/{providerId}/domains/{fqdn}:

    get:
      operationId: get_provider_domain
      description: Retrieve a domain from the given Provider.
      parameters:
      - description: Provider identifier
        in: path
        name: providerId
        required: true
        schema:
          type: string
      - description: Fully qualified domain name
        in: path
        name: fqdn
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Provider'
        "400":
          description: Provider error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Create a domain on the Provider.
      tags:
        - providers
...
components:
  securitySchemes:

    ApiKeyAuth:
      description: Description for what is this security definition being used
      in: header
      name: Authorization
      type: apiKey

    BasicAuth:
      type: http
      scheme: basic
...